JUnit test for account management example TO DO: add exceptions, test massage,test method names and logs.
